Physics colliders logic in progress.

+++ Assets/Scripts/Configs/Game/Colliders/BoxColliderComponentConfig.cs
@@ -5,6 +5,6 @@ namespace Configs.Game.Colliders
     [CreateAssetMenu(fileName = "BoxCollider", menuName = "Configs/Colliders/Box Collider")]
     public class BoxColliderComponentConfig : ColliderComponentConfig
     {
-        public int InnerRadius;
+        public double InnerRadius;
     }
 }
