+++ Assets/Scripts/Configs/Game/Colliders/CircleColliderComponentConfig.cs
@@ -5,6 +5,6 @@ namespace Configs.Game.Colliders
     [CreateAssetMenu(fileName = "CircleCollider", menuName = "Configs/Colliders/Circle Collider")]
     public class CircleColliderComponentConfig : ColliderComponentConfig
     {
-        public int Radius;
+        public double Radius;
     }
 }
